third_party/WebKit/LayoutTests/crypto/aes-cbc-generateKey-parseAlgorithm-failures-expected.txt

CONSOLE ERROR: Invalid keyUsages argument
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Missing required property
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Outside of numeric range
Tests generateKey() with bad AES-CBC parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error is: TypeError: The 3rd argument is neither an array, nor does it have indexed properties.
PASS error is: TypeError: The 3rd argument is neither an array, nor does it have indexed properties.
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
